gateway/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-telegram-bot>=21.0
PyNaCl>=1.5.0
